+++ src/main/java/com/example/Project/ECommerce/Utility/CategoryMetadataFieldValuesId.java
@@ -7,32 +7,33 @@ import java.util.Objects;
 @Embeddable
 public class CategoryMetadataFieldValuesId implements Serializable {
 
-    private int category_id;
-    private int category_metadata_field_id;
+    private long category_id;
+    private long category_metadata_field_id;
 
     public CategoryMetadataFieldValuesId() {
     }
 
-    public CategoryMetadataFieldValuesId(int category_id, int category_metadata_field_id) {
+    public CategoryMetadataFieldValuesId(long category_id, long category_metadata_field_id) {
         this.category_id = category_id;
         this.category_metadata_field_id = category_metadata_field_id;
     }
 
-    public int getCategory_id() {
+    public long getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(int category_id) {
+    public void setCategory_id(long category_id) {
         this.category_id = category_id;
     }
 
-    public int getCategory_metadata_field_id() {
+    public long getCategory_metadata_field_id() {
         return category_metadata_field_id;
     }
 
-    public void setCategory_metadata_field_id(int category_metadata_field_id) {
+    public void setCategory_metadata_field_id(long category_metadata_field_id) {
         this.category_metadata_field_id = category_metadata_field_id;
     }
+
 
     @Override
     public boolean equals(Object o) {
